+++ tilebox-datasets/tilebox/datasets/__init__.py
@@ -3,10 +3,12 @@ import sys
 
 from loguru import logger
 
+# only here for backwards compatibility, to preserve backwards compatibility with older imports
+from tilebox.datasets.aio.timeseries import TimeseriesCollection, TimeseriesDataset
 from tilebox.datasets.sync.client import Client
-from tilebox.datasets.sync.timeseries import TimeseriesCollection, TimeseriesDataset
+from tilebox.datasets.sync.dataset import CollectionClient, DatasetClient
 
-__all__ = ["Client", "TimeseriesCollection", "TimeseriesDataset"]
+__all__ = ["Client", "CollectionClient", "DatasetClient", "TimeseriesCollection", "TimeseriesDataset"]
 
 
 def _init_logging(level: str = "INFO") -> None: