+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.35.0] - 2025-04-29
+
 ## Added
 
 - Support for ingestion of enum values into datasets
+- Pagination support for ingestion / deletion of large numbers of datapoints
+
+## Changed
+
+- Renamed `TimeseriesDataset` to `DatasetClient`
+- Renamed `TimeseriesCollection` to `CollectionClient`
 
 ## [0.34.0] - 2025-04-15
 
@@ -153,7 +161,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.34.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.35.0...HEAD
+[0.35.0]: https://github.com/tilebox/tilebox-python/compare/v0.34.0...v0.35.0
 [0.34.0]: https://github.com/tilebox/tilebox-python/compare/v0.33.1...v0.34.0
 [0.33.1]: https://github.com/tilebox/tilebox-python/compare/v0.33.0...v0.33.1
 [0.33.0]: https://github.com/tilebox/tilebox-python/compare/v0.32.1...v0.33.0